Move math to a utils file instead of in mod.
